tests/submission/PhilipTaylor/canvas/2d.text.font.parse.complex.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.text.font.parse.complex</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.text.html">text</a>.<a href="index.2d.text.font.html">font</a>.<a href="index.2d.text.font.parse.html">parse</a>.complex</h1>
<p class="desc"></p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.text.font.parse">2d.text.font.parse</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.2d.text.font.get">2d.text.font.get</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.2d.text.font.lineheight">2d.text.font.lineheight</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

ctx.font = 'small-caps italic 400 12px/2 Unknown Font, sans-serif';
_assertSame(ctx.font, 'italic small-caps 12px "Unknown Font", sans-serif', "ctx.font", "'italic small-caps 12px \"Unknown Font\", sans-serif'");


});
</script>

<div id="log"></div>
</body></html>
